Add first Robolang interpreter.
This commit adds the implementation of the first interpreter of
Robolang in language/, which supports simple actions and bounded
loops (with a constant number of iterations).

A simple testing framework was also added in test/run_tests.js.
The first test using it is language/robolang/test/test_robolang.js.