Ensure animations are rendered in their final state.
This eliminates the effect of small pauses in the final state
of the animation. A glitch could cause an element not to be
updated in the final state of the animation because it could go
from active (before its end_time) to inactive (after its end_time)
without ever being updated at its end_time.