Lesson 1: fix animation when character leaves grid or is hit.
When the character was being hit, the animation would go "out of sync"
because the animation was timed so that all the asteroids would hit the
ground by the time the animation ends (won't always happen in the game
itself). This was what caused asteroids to move faster when the player
died.

This commit also handles more gracefully the case in which the character
leaves the grid, by showing a different error message.