Store resources in ResourceLoader instead of relying on browser caching (#41)
Before, ResourceLoader loaded the required resources once, in the hope
that when they were loaded a second time (in ElementFactories, for
instance) they would be already cached. This is not always the case.

The fix was to load and store resources in the ResourceLoader. This
module now acts as a "singleton", holding all resources needed at the
time. Element factories directly get resources from it, assuming they
have already been loaded. Modules that require resources have a
"populateResourceLoader" function that is called before the module is
used. The lessons now start only after the resource loader module was
populated and finishes loading everything.