Add Lesson04.
Add game in which a mechanical arm has to collect material from sources
and deposit them into their correct deposits.

It is possible to use machines to transform a bundle of items into
another item, that can further be collected and deposited to another
deposit or machine that accepts that same item.