Update layout (#64)
Put the game canvas in the top left of the window. Success messages, error
messages and instructions appear below the game canvas. Code editor is on
the right, with more space. And commands reference now has a fixed size.

Changed the font to Open Sans, putting the files in the archives.

Unified all css styles in one file called main.css.