Add option to minify bundle.js during build.
Add a --release option to minify the generated bundle.js
in a file called 'bundle-min.js'. This file is still not used anywhere;
however, we may use it in the 'real' client of the app (i.e. not in
tests/lessons/..., which we can keep for development/debugging).

This reduces the size of the file from
1.5M (328K gzipped) to
884K (192K gzipped).