Major architecture design overhaul.
This introduces an architecture that is much more decoupled than what we
had before. It clearly separates what is a game, a challenge and a lesson.
In the near future, the notion of a "course" will also fit in naturally.

It is based on the following concepts:

- A Game is a set of rules in which challenges can be solved.
-- Games are played by a GameRunner (game/game.js).
- A challenge is a specific situation in which a specific game is played.
-- Challenges are represented by a Challenge (game/game.js)
- A Lesson is a sequence of challenges (lesson/lesson.js).
-- Lessons can contain challenges from multiple games (for example a
   "review" lesson). Also, a challenge can be present in multiple
   lessons (i.e. in an introductory lesson, and then again in a
   "review" lesson).

Soon, a course will be defined as a sequence of lessons.

All games, challenges and lessons have an unique identifier (a string).
Challenges are related to games, and lessons to challenges, by their
IDs.

There are also global registers that allow finding registered Challenges
and Lessons, or creating GameRunners for registered games, by their IDs.